~Auto: Commit

+++ src/Tbx/Console/SiteBackup.php
@@ -41,6 +41,8 @@ class SiteBackup extends Iface
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
+        set_time_limit(0);
+
         $this->dataMode = $input->getOption('data');
         $this->site     = $input->getArgument('site');
         $this->destPath = rtrim($input->getArgument('destPath'), '/\\');
@@ -80,7 +82,7 @@ class SiteBackup extends Iface
 
         if (!$this->postRequest($url, $site['secret'], $filename, !$input->getOption('noverify'))) {
             $this->writeError("Error downloading from mirror site");
-            unlink($filename);
+            if (is_file($filename)) unlink($filename);
             return Command::FAILURE;
         }
 
@@ -111,6 +113,8 @@ class SiteBackup extends Iface
             CURLOPT_RETURNTRANSFER => true,
             CURLOPT_POSTFIELDS     => $query,
             CURLOPT_FILE           => $fp,
+            // CURLOPT_CONNECTTIMEOUT => 60,
+            //CURLOPT_TIMEOUT        => 60*60,
             CURLOPT_HTTPHEADER     => [
                 "authorization-key: " . $secret,
             ],
@@ -135,5 +139,110 @@ class SiteBackup extends Iface
 
         return true;
     }
+
+    // Using a background task, does not work for cpanel...
+//    protected function postRequest(Uri|string $srcUrl, string $secret, string $filename, bool $verifyssl = true): bool
+//    {
+//        if (empty($secret)) {
+//            Log::error("Invalid API secret");
+//            return false;
+//        }
+//
+//        $srcUrl = Uri::create($srcUrl)->withScheme('https');
+//        $procUrl = Uri::create($srcUrl)->withScheme('https');
+//
+//        $query = $srcUrl->getQuery();
+//        $srcUrl->reset();
+//
+//        // common opts
+//        $opts = [
+//            CURLOPT_CUSTOMREQUEST  => 'POST',
+//            CURLOPT_RETURNTRANSFER => true,
+//            CURLOPT_POSTFIELDS     => $query,
+//            // CURLOPT_CONNECTTIMEOUT => 60,
+//            //CURLOPT_TIMEOUT        => 0,
+//            CURLOPT_HTTPHEADER     => [
+//                "authorization-key: " . $secret,
+//            ],
+//            CURLOPT_USERAGENT      => Uri::USERAGENT,
+//        ];
+//        if (!$verifyssl) {
+//            $opts[CURLOPT_SSL_VERIFYHOST] = false;
+//            $opts[CURLOPT_SSL_VERIFYPEER] = false;
+//        }
+//
+//        // initiate package creation
+//        Log::debug("Requesting package");
+//        $curl = curl_init($srcUrl->toString());
+//        if ($curl === false) {
+//            Log::error("Cannot open Url: $srcUrl");
+//            return false;
+//        }
+//        curl_setopt_array($curl, $opts);
+//        $res = json_decode(curl_exec($curl));
+//        if(curl_error($curl) || curl_getinfo($curl, CURLINFO_RESPONSE_CODE) != 200) {
+//            Log::error("Request Error: " . $srcUrl);
+//            Log::error("CURL Error: " . curl_error($curl));
+//            Log::error("Request Response: " . curl_getinfo($curl, CURLINFO_RESPONSE_CODE));
+//            return false;
+//        }
+//        curl_close($curl);
+//        if (empty($res->pid) || empty($res->filename)) {
+//            Log::error("Invalid mirror response. No PID returned.");
+//            return false;
+//        }
+//
+//        $pid = $res->pid;
+//        $destFile = $res->filename;
+//
+//        $procUrl->set('pid', $pid);
+//        $procUrl->set('filename', $destFile);
+//        $opts[CURLOPT_POSTFIELDS] = $procUrl->getQuery();
+//        $procUrl->reset();
+//
+//        Log::debug("Downloading package");
+//        $this->write("\n");
+//        while(true) {
+//            $this->write(".");
+//            $fp = fopen($filename, "w");
+//            if ($fp === false) {
+//                Log::error("Cannot open filename: $filename");
+//                return false;
+//            }
+//            $curl = curl_init($procUrl->toString());
+//            if ($curl === false) {
+//                Log::error("Cannot open Url: $procUrl");
+//                return false;
+//            }
+//            $opts[CURLOPT_FILE] = $fp;
+//            curl_setopt_array($curl, $opts);
+//            curl_exec($curl);
+//            if(curl_error($curl) || curl_getinfo($curl, CURLINFO_RESPONSE_CODE) != 200) {
+//                Log::error("Download Request Error: " . curl_error($curl));
+//                Log::error("Download Request Response: " . curl_getinfo($curl, CURLINFO_RESPONSE_CODE));
+//                return false;
+//            }
+//            curl_close($curl);
+//            fclose($fp);
+//
+//            $fp = fopen($filename, "r");
+//            $res = fread($fp, 1024);
+//            fclose($fp);
+//
+//            // this should signal the download is complete
+//            if (json_validate($res)) {
+//                unlink($filename);
+//                $res = json_decode($res);
+//                if ($res->complete ?? false) {
+//                    Log::error("Error downloading package: {$filename}");
+//                    return false;
+//                }
+//            } else {
+//                Log::debug("Download completed: {$filename}");
+//                return true;
+//            }
+//            sleep(10);
+//        }
+//    }
 
 }
